Fix destination of COPY user defined config
Regarding to https://docs.docker.com/reference/builder/#copy for COPY command "[...]If <dest> does not end with a trailing slash, it will be considered a regular file and the contents of <src> will be written at <dest>.[...]". For this reason it is required to add a trailing slash so user defined /config/ will be added.