Make entry point more versatile for USER.
The user and group can now be choosen using the `RUN_AS` environment
variable. `RUN_AS` should follow the `user[:group]` pattern and both
user/group numbers and user/group names are accepted.

The `chown`/`gosu` mechanism is skipped when the user is either `0`
or `root` or if the command is not `elasticsearch`.